requirements.txt
libvirt-python
flask
